Continue refactoring work
